fix: make middleware working when connecting from ts
